refactor(symlink): split symlink handler tests into per-area files
handlers/symlink.rs (1663 LOC: 365 code, 837 tests) is promoted to a
directory and the tests fan out by topic:

- symlink/mod.rs              — handler code only (~615 LOC), declares
                                 `#[cfg(test)] mod tests;`
- symlink/tests/mod.rs        — shared fixtures (test_pather,
                                 default_config) + small inline tests
                                 (default rule, pack-prefix
                                 interaction, protected paths,
                                 force_home matching)
- symlink/tests/cascade.rs    — Priority-tier resolution: force_home,
                                 _home/_xdg, _app/, _lib/, force_app,
                                 app_aliases, plus the `home.` /
                                 `app.` / `xdg.` / `lib.` file prefixes
- symlink/tests/integration.rs — Routing-override conflicts, custom
                                 [symlink.targets], wholesale-vs-per-
                                 file behaviour, _lib/ warnings

Shared fixtures are `pub(super)`; per-area files import them via
`use super::{...}`. No production change; same 1140 tests still pass.

Test setup duplication audit and consolidation is left as follow-up
work — the priority here was getting the giant file split into
focused topical files.
